add tools and refactor
