+++ tools/getCoverage.sh
@@ -1,0 +1,5 @@
+#!/bin/sh
+
+cd nn
+go test -coverprofile=profile ./...
+go tool cover -html=profile
